+++ src/flamenco/vm/fd_vm_disasm.c
@@ -146,7 +146,7 @@ fd_vm_disasm_instr_jmp( fd_sbpf_instr_t            instr,
   if( FD_UNLIKELY( instr.opcode.normal.op_mode==FD_SBPF_OPCODE_JMP_OP_MODE_CALL ) ) {
     switch ( instr.opcode.normal.op_src ) {
     case FD_SBPF_OPCODE_SOURCE_MODE_IMM: {
-      fd_sbpf_syscalls_t const * syscall = syscalls ? fd_sbpf_syscalls_query_const( syscalls, instr.imm, NULL ) : NULL;
+      fd_sbpf_syscalls_t const * syscall = syscalls ? fd_sbpf_syscalls_query_const( syscalls, (ulong)instr.imm, NULL ) : NULL;
       if( syscall ) { /* FIXME: THESE CODE PATHS CURRENTLY NOT EXERCISED BY UNIT TEST */
         char const * name = syscall->name;
         if( name ) OUT_PRINTF( "syscall%s %s",     suffix, name      );
